requirements.txt
flask>=2.0,<4.0
pyyaml>=5.4
